refactor(server): remove workspace flat map cache service
