+++ packages/twenty-server/src/engine/metadata-modules/flat-role-target/services/workspace-flat-role-target-map-cache.service.ts
@@ -29,6 +29,7 @@ export class WorkspaceFlatRoleTargetMapCacheService extends WorkspaceCacheProvid
   }: {
     workspaceId: string;
   }) {
+    // TODO
     await Promise.all([
       this.workspaceFlatRoleTargetByAgentIdService.invalidateCache({
         workspaceId,
