--- packages/twenty-server/src/engine/workspace-flat-map-cache/decorators/workspace-flat-map-cache.decorator.ts
@@ -1,11 +1,0 @@
-import { SetMetadata } from '@nestjs/common';
-
-import { type AllMetadataName } from 'twenty-shared/metadata';
-
-import { type MetadataToFlatEntityMapsKey } from 'src/engine/metadata-modules/flat-entity/types/metadata-to-flat-entity-maps-key';
-
-export const WORKSPACE_FLAT_MAP_CACHE_KEY = 'workspaceFlatMapCacheKey';
-
-export const WorkspaceFlatMapCache = (
-  cacheKey: MetadataToFlatEntityMapsKey<AllMetadataName>,
-) => SetMetadata(WORKSPACE_FLAT_MAP_CACHE_KEY, cacheKey);
