--- packages/twenty-server/src/engine/workspace-flat-map-cache/exceptions/workspace-flat-map-cache.exception.ts
@@ -1,12 +1,0 @@
-import {
-  appendCommonExceptionCode,
-  CustomException,
-} from 'src/utils/custom-exception';
-
-export class WorkspaceFlatMapCacheException extends CustomException<
-  keyof typeof WorkspaceFlatMapCacheExceptionCode
-> {}
-
-export const WorkspaceFlatMapCacheExceptionCode = appendCommonExceptionCode({
-  MISSING_DECORATOR: 'MISSING_DECORATOR',
-} as const);
